Fix functions that calculate pass rate per instructor

+++ lab3_files/src/readAndWrite.h
@@ -5,6 +5,6 @@
 std::vector<std::vector<std::string>> readData(std::string fname);
 
 // printMap() function to print the contents of a map
-void printMap(FILE *fp, std::map<std::string, int> myMap);
+void printMap(FILE *fp, std::map<std::string, double> myMap);
 
 #endif
